# microservices-on-openshift/quarkus/product/product_openapi3.yaml
...
openapi: 3.0.1
info:
  title: Product API
  contact:
    name: Product API Support
    url: http://amazin.com/contact
    email: techsupport@amazin.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.1
...
paths:
...
  /products/type/{type}:
    get:
      parameters:
      - name: type
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
...
  /products/{productId}:
    get:
      parameters:
      - name: productId
        in: path
        required: true
        schema:
          format: int32
          type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
components:
  schemas:
    ProductType:
      enum:
      - CLOTHES
      - FOOD
      - GADGETS
      type: string
    Product:
      type: object
      properties:
        basketIndex:
          format: int32
          type: integer
        id:
          type: string
        name:
          type: string
        price:
          format: float
          type: number
        type:
          $ref: '#/components/schemas/ProductType'